refac

+++ code/check_sum.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint32_t checksum_oper_calc(uint8_t *buf, uint32_t size);
+uint8_t checksum_oper_calc(uint8_t *buf, uint32_t size);
 uint32_t is_valid_checksum_oper(uint8_t *buf, uint32_t size);
 uint32_t is_valid_checksum_kama(uint8_t *buf, uint32_t size);
